refactor: kakaoMap 의존성 제거
