+++ app/src/main/java/com/bestapp/zipbab/App.kt
@@ -3,10 +3,7 @@ package com.bestapp.zipbab
 import android.app.Application
 import com.google.firebase.FirebaseApp
 import com.google.firebase.perf.metrics.AddTrace
-import com.kakao.vectormap.KakaoMapSdk
 import dagger.hilt.android.HiltAndroidApp
-
-
 
 @HiltAndroidApp
 class App : Application() {
@@ -15,6 +12,5 @@ class App : Application() {
     override fun onCreate() {
         super.onCreate()
         FirebaseApp.initializeApp(applicationContext)
-        KakaoMapSdk.init(this, BuildConfig.KAKAO_NATIVE_KEY)
     }
 }
