+++ app/src/main/java/com/bestapp/zipbab/ui/meetupmap/NaverMapExtension.kt
@@ -1,0 +1,20 @@
+package com.bestapp.zipbab.ui.meetupmap
+
+import kotlin.math.atan2
+import kotlin.math.cos
+import kotlin.math.sin
+import kotlin.math.sqrt
+
+//fun haversine(latlng1: LatLng, latlng2: LatLng) =
+//    haversine(latlng1.latitude, latlng1.longitude, latlng2.latitude, latlng2.longitude)
+
+fun haversine(lat1: Double, lon1: Double, lat2: Double, lon2: Double): Double {
+    val R = 6371.0 // 지구 반지름 (킬로미터)
+    val dLat = Math.toRadians(lat2 - lat1)
+    val dLon = Math.toRadians(lon2 - lon1)
+    val a = sin(dLat / 2) * sin(dLat / 2) +
+            cos(Math.toRadians(lat1)) * cos(Math.toRadians(lat2)) *
+            sin(dLon / 2) * sin(dLon / 2)
+    val c = 2 * atan2(sqrt(a), sqrt(1 - a))
+    return R * c // 거리 (킬로미터)
+}
