+++ app/src/main/res/layout/fragment_meet_up_map.xml
@@ -8,14 +8,14 @@
     android:layout_height="match_parent"
     tools:context=".ui.meetupmap.MeetUpMapFragment">
 
-    <com.kakao.vectormap.MapView
-        android:id="@+id/mv"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+<!--    <com.kakao.vectormap.MapView-->
+<!--        android:id="@+id/mv"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="match_parent"-->
+<!--        app:layout_constraintBottom_toBottomOf="parent"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toTopOf="parent" />-->
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/fab_gps"
