+++ settings.gradle.kts
@@ -16,9 +16,6 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        maven {
-            url = java.net.URI("https://devrepo.kakao.com/nexus/content/groups/public/")
-        }
     }
 }
